+++ pysys-examples/internal/testcases/PySys_internal_068/Input/NestedNotVerified/run.py
@@ -1,0 +1,11 @@
+# -*- coding: latin-1 -*-
+
+from pysys.constants import *
+from pysys.basetest import BaseTest
+from pysys.exceptions import *
+
+class PySysTest(BaseTest):
+	def execute(self):
+		pass
+	def validate(self):
+		pass